fixers/doc-drift-fixer/requirements.txt
langchain
langchain-openai
gitpython
pygithub
rich
pytest
python-dotenv
requests
